fix: skip new issue after updating

+++ dev/io.openliberty.data.internal_fat_jpa/test-applications/DataJPATestApp/src/test/jakarta/data/jpa/web/DataJPATestServlet.java
@@ -201,6 +201,26 @@ public class DataJPATestServlet extends FATServlet {
         return isHibernate();
     }
 
+    /**
+     * Temporary method to allow skipping tests for tests that
+     * fail due to incompatibilities between our Jakarta Data provider
+     * and Hibernate's Jakarta Persistence provider on a specific database.
+     *
+     * @param driver - driver name prefix (i.e. derby)
+     * @param issues - the issues that describe why the test must be skipped on Hibernate
+     * @return boolean - true if we need to skip the test, false otherwise.
+     */
+    public static boolean skipForHibernateByDatabase(String driver, String... issues) {
+        boolean isHibernateAndDatabase = isHibernate() && System.getenv("DB_DRIVER").contains(driver);
+
+        if (isHibernateAndDatabase) {
+            System.out.println("Skipping test because database is " + System.getenv("DB_DRIVER") + " and " + Arrays.asList(issues));
+        }
+
+        return isHibernateAndDatabase;
+
+    }
+
     @Override
     public void init(ServletConfig config) throws ServletException {
         // Some read-only data that is prepopulated for tests:
@@ -1169,6 +1189,10 @@ public class DataJPATestServlet extends FATServlet {
         String jdbcJarName = System.getenv().getOrDefault("DB_DRIVER", "UNKNOWN");
         boolean supportsOrderByForUpdate = !jdbcJarName.startsWith("derby");
 
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         cities.save(new City("Milwaukee", "Wisconsin", 577222, Set.of(414)));
         cities.save(new City("Green Bay", "Wisconsin", 107395, Set.of(920)));
         cities.save(new City("Superior", "Wisconsin", 26751, Set.of(534, 715)));
@@ -1292,7 +1316,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testFindAndDeleteReturningIdClassList(HttpServletRequest request, HttpServletResponse response) {
-
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
         cities.save(new City("Davenport", "Iowa", 101724, Set.of(563)));
         cities.save(new City("Sioux City", "Iowa", 85797, Set.of(712)));
         cities.save(new City("Iowa City", "Iowa", 74828, Set.of(319)));
@@ -2402,7 +2428,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testFindAndDeleteInTransaction() throws Exception {
-
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
         cities.save(new City("Grand Forks", "North Dakota", 59845, Set.of(701)));
         cities.save(new City("Bismarck", "North Dakota", 77772, Set.of(701)));
         cities.save(new City("Fargo", "North Dakota", 136285, Set.of(701)));
@@ -3760,6 +3788,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testMultipleThreadsVersionedUpdate() throws Exception {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot update with pessimistic lock
+        }
         orders.deleteAll();
 
         PurchaseOrder o = new PurchaseOrder();
@@ -3858,6 +3889,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testOneToOne() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
 
         drivers.deleteByFullNameEndsWith(" TestOneToOne");
 
@@ -4161,7 +4195,11 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testSaveInTransaction() throws Exception {
-        cities.deleteByStateName("Montana");
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            cities.deleteVoidByStateName("Montana");; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        } else {
+            cities.deleteByStateName("Montana");
+        }
 
         tran.begin();
         try {
@@ -4186,7 +4224,11 @@ public class DataJPATestServlet extends FATServlet {
         assertEquals(117116, city.population);
         assertEquals(Set.of(406), city.areaCodes);
 
-        cities.deleteByStateName("Montana");
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            cities.deleteVoidByStateName("Montana");; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        } else {
+            cities.deleteByStateName("Montana");
+        }
     }
 
     /**
@@ -4769,6 +4811,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testUpdateEntityWithIdClassAndVersion() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot update with pessimistic lock
+        }
 
         CityId mnId = CityId.of("Rochester", "Minnesota");
         CityId nyId = CityId.of("Rochester", "New York");
@@ -4820,7 +4865,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testUpdateInTransaction() throws Exception {
-
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot update with pessimistic lock
+        }
         Mobile m1 = mobilePhones.insert(Mobile.of(OS.ANDROID,
                                                   List.of("settings",
                                                           "messages",
@@ -5109,6 +5156,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testVersionedDelete() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
         orders.deleteAll();
 
         PurchaseOrder o1 = new PurchaseOrder();
@@ -5228,6 +5278,9 @@ public class DataJPATestServlet extends FATServlet {
      */
     @Test
     public void testVersionedUpdate() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot update with pessimistic lock
+        }
         orders.deleteAll();
 
         PurchaseOrder o1 = new PurchaseOrder();
